separated base_model from base
