+++ api/v1/models/profile.py
@@ -14,7 +14,8 @@ from sqlalchemy import (
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
 from datetime import datetime
-from api.v1.models.base import Base, BaseModel
+from api.v1.models.base import Base
+from api.v1.models.base_model import BaseModel
 from uuid_extensions import uuid7
